+++ src/oci/object_storage/transfer/upload_manager.py
@@ -27,8 +27,9 @@ class UploadManager:
         An advantage of using multi-part uploads is the ability to retry individual failed parts, as well as being
         able to upload parts in parallel to reduce upload time.
 
-        PLEASE NOTE that the operations are NOT thread-safe, and you should provide the UploadManager class
-        with its own Object Storage client that isn't used elsewhere.
+        PLEASE NOTE that the timeout for the object storage client is overwritten to `None` for all operations which
+        call object storage. For this reason, the operations are NOT thread-safe, and you should provide the
+        UploadManager class with its own Object Storage client that isn't used elsewhere.
         For more information please see `Known Issues <https://oracle-cloud-infrastructure-python-sdk.readthedocs.io/en/latest/known-issues.html>`_
 
         :param ObjectStorageClient object_storage_client: